Finalização da parte relativa ao JDBC - Todos os métodos da facade estão funcionando

+++ SistemaAcompanhamentoDefeitos/src/persistencia/EntradaDAO_JDBC.java
@@ -3,6 +3,9 @@
 package persistencia;
 
 import java.sql.ResultSet;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 import logica.Entrada;
 
 public class EntradaDAO_JDBC implements EntradaDAO {
@@ -14,15 +17,49 @@ public class EntradaDAO_JDBC implements EntradaDAO {
 
 		try {
 			banco.abreConexao();
-			sql = "INSERT INTO entrada (cod_entrada, data_postada, desc_entrada, status_problema, cod_historico) values ('"
-					+ entrada.getCodEntrada() + "','" + entrada.getDataPostada() + "','" + entrada.getDescricao()
-					+ "','" + entrada.getStatusProblema() + "','" + entrada.getCodHistorico().getCodHistorico() + "')";
-			banco.update(sql);
 
-			sql = "INSERT INTO entrada_funcionario (cod_entrada, matricula_func) values ('" + entrada.getCodEntrada()
-					+ "','" + entrada.getMatriculaFunc().getMatricula() + "')";
-			banco.update(sql);
-			return true;
+			sql = "SELECT * FROM entrada WHERE desc_entrada='" + entrada.getDescricao() + "' and data_postada='"
+					+ formataData(entrada.getDataPostada()) + "'";
+			ResultSet rs = banco.query(sql);
+
+			if (!rs.next()) {
+				int codHistorico = 0;
+				sql = "SELECT * FROM historico_entrada WHERE cod_problema='"
+						+ entrada.getHistorico().getProblema().getCodProblema() + "'";
+				rs = banco.query(sql);
+
+				if (rs.next()) {
+					codHistorico = rs.getInt("cod_historico");
+
+					sql = "SELECT * FROM funcionario WHERE matricula='" + entrada.getFuncionario().getMatricula() + "'";
+					rs = banco.query(sql);
+
+					if (rs.next()) {
+						sql = "INSERT INTO entrada (data_postada, desc_entrada, status_problema, cod_historico, matricula_func) values ('"
+								+ formataData(entrada.getDataPostada()) + "','" + entrada.getDescricao() + "','"
+								+ entrada.getStatusProblema() + "','" + codHistorico + "','"
+								+ entrada.getFuncionario().getMatricula() + "')";
+						banco.update(sql);
+
+						if (entrada.getStatusProblema().equalsIgnoreCase("encerrado") || entrada.getStatusProblema().equalsIgnoreCase("finalizado") || entrada.getStatusProblema().equalsIgnoreCase("concluído")) {
+							
+							sql = "UPDATE problema SET data_resolucao='"+formataData(entrada.getDataPostada())+"' WHERE cod_problema='"+entrada.getHistorico().getProblema().getCodProblema()+"'";
+							banco.update(sql);
+							
+						}
+											
+						return true;
+
+					} else {
+						return false;
+					}
+				} else {
+					return false;
+				}
+			}
+			else {
+				return false;
+			}
 		}
 
 		catch (Exception e) {
@@ -40,9 +77,18 @@ public class EntradaDAO_JDBC implements EntradaDAO {
 
 		try {
 			banco.abreConexao();
-			sql = "DELETE FROM entrada WHERE cod_entrada='" + codEntrada + "'";
-			banco.update(sql);
-			return true;
+			sql = "SELECT * FROM entrada WHERE cod_entrada='" + codEntrada + "'";
+			ResultSet rs = banco.query(sql);
+
+			if (rs.next()) {
+				sql = "DELETE FROM entrada WHERE cod_entrada='" + codEntrada + "'";
+				banco.update(sql);
+				return true;
+			}
+
+			else {
+				return false;
+			}
 		}
 
 		catch (Exception e) {
@@ -57,7 +103,10 @@ public class EntradaDAO_JDBC implements EntradaDAO {
 	}
 
 	public Entrada buscarEntrada(int codEntrada) {
+		FuncionarioDAO_JDBC fDAO = new FuncionarioDAO_JDBC();
+
 		Entrada entrada = new Entrada();
+
 		try {
 			banco.abreConexao();
 			sql = "SELECT * FROM entrada WHERE cod_entrada='" + codEntrada + "'";
@@ -68,14 +117,11 @@ public class EntradaDAO_JDBC implements EntradaDAO {
 				entrada.setDataPostada(rs.getDate("data_postada"));
 				entrada.setDescricao(rs.getString("desc_entrada"));
 				entrada.setStatusProblema(rs.getString("status_problema"));
-				entrada.getCodHistorico().setCodHistorico(rs.getInt("cod_historico"));
+				entrada.setFuncionario(fDAO.buscarFuncionario(rs.getInt("matricula_func")));
 
-				String sql2 = "SELECT matricula_func FROM entrada_funcionario WHERE cod_entrada='" + codEntrada + "'";
-				ResultSet rs2 = banco.query(sql2);
+				sql = "SELECT";
+				entrada.getHistorico().setCodHistorico(rs.getInt("cod_historico"));
 
-				if (rs2.next()) {
-					entrada.getMatriculaFunc().setMatricula(rs.getInt("matricula_func"));
-				}
 			}
 
 			return entrada;
@@ -90,6 +136,12 @@ public class EntradaDAO_JDBC implements EntradaDAO {
 		finally {
 			banco.fechaConexao();
 		}
+	}
+
+	private String formataData(Date date) {
+		SimpleDateFormat format = new SimpleDateFormat("dd/MM/yyyy");
+		String data = format.format(date);
+		return data;
 	}
 
 }
